--- back_end/src/utils/web_crawler/Warren.json
@@ -1,35 +1,0 @@
-{
-    "Warren Writing A": ["WCWP 10A"],
-    "Warren Writing B": ["WCWP 10B"],
-    "Ethics and Society": [
-        ["POLI 27", "PHIL 27"],
-        ["POLI 28", "PHIL 28"]
-    ],
-    "Formal Skills":["CSE 3", "MATH 3C", "MATH 10A", "MATH 4C", "MATH 20A", "MATH 10B", "MATH 20B", "MATH 11", "MATH 18", "PHIL 10", "PHIL 12", "POLI 30", "POLI 30D", "POLI 5", "POLI 5D", "Econ 5", "PSYC 60", "PSYC 70", "SOCI 60", "LIGN 17", "COGS 3", "COGS 8", "COGS 14A", "COGS 14B", "COGS 18", "CSE 7",  "CSE 8A", "CSE 8B", "CSE 11", "CSS 1", "CSS 2", "ECE 5", "ECE 15", "ECE 25", "MAE 5", "NANO 15", "CENG 15","HDS 60"],
-    "AREA STUDY": {
-        "HUMANITIES": ["HILD 2A", "HIUS 103", "HIUS 112"],
-        "SOCIAL SCIENCE":["POLI 10", "POLI 100A", "POLI 100B"]
-    },
-    "program of concentration":[
-        ["HIEU 106", "HILA 110", "HINE 145", "HISC 102", "HISC 105"],
-        ["HILD 2A", "HILD 2B", "HILD 2C"],
-        ["PHIL 13", "PHIL 16", "PHIL 26", "PHIL 27", "PHIL 28"],
-        ["PHIL 139", "PHIL 147"],
-        ["MUS 5", "MUS 7", "MUS 8", "MUS 11", "MUS 13", "MUS 14", "MUS 15", "MUS 20"],
-        ["MUS 107", "MUS 110", "MUS 111", "MUS 114", "MUS 115", "MUS 116", "MUS 150", "MUS 151", "MUS 176"],
-        ["ENVR 30"],
-        ["ENVR 110"],
-        ["ECON 131", "ECON 132", "ECON 145"],
-        ["POLI 140E", "POLI 125A", "POLI 162"],
-        ["USP 2", "USP 124", "USP 144", "USP 171"],
-        ["SOCI 149"]
-    ],
-    "area of studies":[
-        ["HILD 2A", "HILD 2B", "HILD 2C"],
-        ["HIEU 106", "HILA 110", "HINE 145", "HISC 102", "HISC 105"],
-        ["PHIL 139", "PHIL 147"],
-        ["ENVR 30"],
-        ["ENVR 110"],
-        ["ECON 131", "ECON 132", "ECON 145"]
-    ]
-}
